Fixes in Scanning module (Issue #66)
In the scanning module, ALLOWED_SCAN_TYPES and SUPPORTED_SCANNING_MODES conflict. Unified on SUPPORTED_SCANNING_MODES.

In the scanning module, the new _frequency_tune method is calling _process_queue, but, with the exception of pass_count, the updated values aren't passed back up to the caller, and interval is never initialized. Could set up a named tuple or something to pass the values back, but seems more straightforward to return the _process_queue call to the main frequency scan method. Added back in interval initialization.

In the rigctl module, several of the "get" methods were looking for either a string or a float, neither of which the _request method can return. It always returns a unicode string. Changed the expected type to basestring in the "get" methods.