Add placeholder Qt window
This window will be built incrementally. For now we just added a placeholder
with a splitter and two labels.

Please note I added a conditional block to enable us to test the qt_ui.py
file directly during development. That will be moved to the main rig-remote.py
as the UI development progresses.